Refactor cancel_game to delete misses

+++ api.py
@@ -358,11 +358,17 @@ class BattleshipAPI(remote.Service):
     def cancel_game(self, request):
         """Cancels an active game"""
         game = get_by_urlsafe(request.url_safe_game_key, Game)
-        game_pieces = Piece.query(Piece.game == game.key).fetch()
-        for piece in game_pieces:
-            piece.key.delete()
-        game.key.delete()
-        return StringMessage(message="Game deleted")
+        if game:
+            game_pieces = Piece.query(Piece.game == game.key).fetch()
+            game_misses = Miss.query(Miss.game == game.key).fetch()
+            for piece in game_pieces:
+                piece.key.delete()
+            for miss in game_misses:
+                miss.key.delete()
+            game.key.delete()
+            return StringMessage(message="Game deleted")
+        else:
+            return StringMessage(message="Game does not exist")
 
 # - - - temp api to place dummy pices in the datastore  - - - - - - - - - - - -
